-Added code to throw and pick up weapons  still pretty early. Can have unknown bugs/unstable

+++ graphics/gfx_core.h
@@ -80,8 +80,8 @@ namespace gfx
 		glm::vec3 camera;
 		glm::vec3 zoom;
 		glm::mat4 MVP;
-		float FrameTime;
-		int SkipFrames;
+		float 	FrameTime;
+		size_t 	SkipFrames;
 
 	protected:
 		///Protected data
